+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 - AMI command results tracking fixed
 - Return AMI command result with multiple events in a Future
+- Return AsyncAGI command result in a Future
 - Add two examples
 - Start internal refactoring